All the image stuff.

+++ src/Command/IngestCommand.php
@@ -5,18 +5,15 @@ namespace App\Command;
 use App\Entity\Episode;
 use App\FlysystemAssetManager\File;
 use App\FlysystemAssetManager\FlysystemAssetManager;
-use App\Form\FileType;
 use App\Podcast\Episodes;
 use App\Repository\EpisodeRepository;
 use Doctrine\Common\Persistence\ObjectManager;
 use GuzzleHttp\Client;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\Form\FormTypeInterface;
 
 class IngestCommand extends Command
 {
@@ -38,26 +35,10 @@ class IngestCommand extends Command
     private $objectManager;
 
     /**
-     * @var StorageInterface
-     */
-    private $storage;
-
-    /**
-     * @var PropertyMappingFactory
-     */
-    private $propertyMappingFactory;
-
-    /**
      * @var FlysystemAssetManager
      */
     private $flysystemAssetManager;
 
-    /**
-     * IngestTestCommand constructor.
-     * @param Episodes $episodes
-     * @param EpisodeRepository $episodeRepository
-     * @param ObjectManager $objectManager
-     */
     public function __construct(Episodes $episodes, EpisodeRepository $episodeRepository, ObjectManager $objectManager, FlysystemAssetManager $flysystemAssetManager)
     {
         $this->episodes = $episodes;
@@ -86,7 +67,7 @@ class IngestCommand extends Command
         $io = new SymfonyStyle($input, $output);
         $force = $input->getOption('force');
 
-        $allExportedEpisodes = collect($this->episodes->getAll())->slice(0, 3);
+        $allExportedEpisodes = collect($this->episodes->getAll())->slice(0, 12);
 
         $allExportedEpisodes->each(function (Episode $exportedEpisode) use ($io, $force) {
             $existingEpisode = $this->episodeRepository->findOneByGuid($exportedEpisode->getGuid());
@@ -97,26 +78,26 @@ class IngestCommand extends Command
 
             if ($existingEpisode) {
                 $io->writeln(sprintf("%s: %s already exists", $existingEpisode->getNumber(), $existingEpisode->getTitle()));
-                if ($existingEpisode->getBackgroundImageUrl()) {
+                if ($existingEpisode->getBackgroundImageUrl() && $this->flysystemAssetManager->exists($existingEpisode->getBackgroundImage())) {
                     $io->writeln(sprintf(" - removing existing background image: %s", $this->flysystemAssetManager->getUrl($existingEpisode->getBackgroundImage())));
                     $this->flysystemAssetManager->delete($existingEpisode->getBackgroundImage());
                     $existingEpisode->setBackgroundImageUrl(null);
                 }
 
                 if ($existingEpisode->getPristineMediaUrl()) {
-                    if ($this->flysystemAssetManager->exists($existingEpisode->getPristineMedia())) {
+                    if ($existingEpisode->getPristineMedia() && $this->flysystemAssetManager->exists($existingEpisode->getPristineMedia())) {
                         $io->writeln(sprintf(" - removing existing pristine media: %s", $this->flysystemAssetManager->getUrl($existingEpisode->getPristineMedia())));
                         $this->flysystemAssetManager->delete($existingEpisode->getPristineMedia());
-                        $existingEpisode->setPristineMediaUrl(null);
                     }
+                    $existingEpisode->setPristineMediaUrl(null);
                 }
 
-                if ($existingEpisode->getMediaUrl()) {
-                    if ($this->flysystemAssetManager->exists($existingEpisode->getMedia())) {
+                if ($existingEpisode->getMediaUrl() && parse_url($existingEpisode->getMediaUrl(), PHP_URL_SCHEME) === 'flysystem') {
+                    if ($existingEpisode->getMedia() && $this->flysystemAssetManager->exists($existingEpisode->getMedia())) {
                         $io->writeln(sprintf(" - removing existing media: %s", $this->flysystemAssetManager->getUrl($existingEpisode->getMedia())));
                         $this->flysystemAssetManager->delete($existingEpisode->getMedia());
-                        $existingEpisode->setMediaUrl(null);
                     }
+                    $existingEpisode->setMediaUrl(null);
                 }
 
                 $existingEpisode->refreshFrom($exportedEpisode);
@@ -147,25 +128,18 @@ class IngestCommand extends Command
                 $response = $client->get($existingEpisode->getMediaUrl());
                 fclose($handle);
 
-                $pristineMediaFile = new File(
+                $pristineMediaFile = File::create(
                     'content',
                     Episode::generatePristineMediaPath($existingEpisode, $originalFileName),
                     $response->getHeader('content-type')[0],
                     (int) $response->getHeader('content-length')[0]
                 );
 
-                $mediaFile = new File(
-                    'content',
-                    Episode::generateMediaPath($existingEpisode, $originalFileName),
-                    $response->getHeader('content-type')[0],
-                    (int) $response->getHeader('content-length')[0]
-                );
-
-                $this->flysystemAssetManager->writeFromFile($pristineMediaFile, $tmpFile);
-                $this->flysystemAssetManager->writeFromFile($mediaFile, $tmpFile);
+                $this->flysystemAssetManager->writeOrUpdateFromFile($pristineMediaFile, $tmpFile);
 
                 $existingEpisode->setPristineMediaUrl($pristineMediaFile->toUrl());
-                $existingEpisode->setMediaUrl($mediaFile->toUrl());
+
+                unlink($tmpFile);
             }
 
             if (substr( $existingEpisode->getBackgroundImageUrl(), 0, 8 ) === "https://") {
@@ -187,17 +161,21 @@ class IngestCommand extends Command
                 $response = $client->get($existingEpisode->getBackgroundImageUrl());
                 fclose($handle);
 
-                $backgroundImageFile = new File(
+                $backgroundImageFile = File::create(
                     'content',
                     Episode::generateBackgroundImagePath($existingEpisode, $originalFileName),
                     $response->getHeader('content-type')[0],
                     (int) $response->getHeader('content-length')[0]
                 );
 
-                $this->flysystemAssetManager->writeFromFile($backgroundImageFile, $tmpFile);
+                $this->flysystemAssetManager->writeOrUpdateFromFile($backgroundImageFile, $tmpFile);
 
                 $existingEpisode->setBackgroundImageUrl($backgroundImageFile->toUrl());
+
+                unlink($tmpFile);
             }
+
+            $existingEpisode->setDuration(null);
 
             $this->objectManager->flush();
         });
